requirements.txt
backports.tarfile==1.2.0
grpcio==1.66.0
importlib-metadata==8.0.0
importlib-resources==6.4.0
ipykernel==6.28.0
jaraco.text==3.12.1
lz4==4.3.3
openai==1.42.0
ordered-set==4.1.0
pinecone-client==5.0.1
pip-chill==1.0.3
protoc-gen-openapiv2==0.0.1
python-dotenv==1.0.1
tomli==2.0.1
